Move masternode ping into own file
